refactor: rename Visitor to ContentVisitor because of new visitors
