+++ src/core/models/character_class.hpp
@@ -8,6 +8,7 @@
 #include <vector>
 
 #include "core/basic_mechanics/dice.hpp"
+#include "core/content_visitors/content_visitor.hpp"
 #include "core/models/effect_holder/feature.hpp"
 #include "core/models/feature_holder.hpp"
 
@@ -34,7 +35,7 @@ public:
      * @brief Accepts a visitor
      * @param visitor pointer to the visitor
      */
-    virtual void accept(Visitor* visitor) const override final;
+    virtual void accept(ContentVisitor* visitor) const override final;
 
     // the type of hit dice for this class
     const Dice hit_dice;
@@ -51,7 +52,7 @@ inline CharacterClass::CharacterClass(
     : FeatureHolder(name, source_file_path, std::move(features)), hit_dice(hit_dice), asi_levels(asi_levels),
       subclass_level(subclass_level) {}
 
-inline void CharacterClass::accept(Visitor* visitor) const { visitor->visit(this); }
+inline void CharacterClass::accept(ContentVisitor* visitor) const { visitor->visit(this); }
 
 } // namespace dnd
