+++ src/core/models/character_subrace.hpp
@@ -7,6 +7,7 @@
 #include <utility>
 #include <vector>
 
+#include "core/content_visitors/content_visitor.hpp"
 #include "core/models/effect_holder/feature.hpp"
 #include "core/models/feature_holder.hpp"
 
@@ -31,7 +32,7 @@ public:
      * @brief Accepts a visitor
      * @param visitor pointer to the visitor
      */
-    virtual void accept(Visitor* visitor) const override final;
+    virtual void accept(ContentVisitor* visitor) const override final;
 
     // the name of the race this is a subrace of
     const std::string race_name;
@@ -43,7 +44,7 @@ inline CharacterSubrace::CharacterSubrace(
 ) noexcept
     : FeatureHolder(name, source_file_path, std::move(features)), race_name(race_name) {}
 
-inline void CharacterSubrace::accept(Visitor* visitor) const { visitor->visit(this); }
+inline void CharacterSubrace::accept(ContentVisitor* visitor) const { visitor->visit(this); }
 
 } // namespace dnd
