+++ src/gui_app/CMakeLists.txt
@@ -4,4 +4,4 @@ target_sources(${DND_GUI_APP}
     gui_launcher.cpp
 )
 
-add_subdirectory(visitors)
+add_subdirectory(content_visitors)
