+++ src/gui_app/content_visitors/list_visitor.cpp
@@ -17,7 +17,7 @@
 #include "core/models/effect_holder/feature.hpp"
 #include "core/models/item.hpp"
 #include "core/models/spell.hpp"
-#include "core/visitors/visitor.hpp"
+#include "core/content_visitors/content_visitor.hpp"
 
 void dnd::ListVisitor::visit(const Character* character_ptr) {
     string_list.emplace_back(fmt::format(
