+++ src/gui_app/content_visitors/list_visitor.hpp
@@ -7,6 +7,7 @@
 #include <utility>
 #include <vector>
 
+#include "core/content_visitors/content_visitor.hpp"
 #include "core/models/character.hpp"
 #include "core/models/character_class.hpp"
 #include "core/models/character_race.hpp"
@@ -16,14 +17,13 @@
 #include "core/models/effect_holder/feature.hpp"
 #include "core/models/item.hpp"
 #include "core/models/spell.hpp"
-#include "core/visitors/visitor.hpp"
 
 namespace dnd {
 
 /**
  * @brief A visitor for creating a list of content pieces (name and type for each)
  */
-class ListVisitor : public Visitor {
+class ListVisitor : public ContentVisitor {
 public:
     /**
      * @brief Return a vector of list-suitable strings
